:sparkles: Add support for multiple features
Add support for EncryptedAttributes in the SAML response.

Allow not setting certfile/keyfile when doing a artifact request. I
addded this so the tests would succeed again. But this might be valid
use case as well.

Add test for using CarriedKeyName when using EncryptedAttributes.

Add support for using CarriedKeyName instead of RetrievalMethod.

Add support for an extended_output mode for 'get_attributes' to allow
the XML attributes of the AttributeValue element being read.

Add a method to directly process a POST ACS response. This is used by Digid-eherkenning.

Add a option/hack to disable Signature Wrapping attack protection, which
is needed when an IDP returns a Advice section with Signatures/Response
data within.

Only sign POST authNRequest if the 'authnRequestsSigned' option is set
in the security configuration.

In testLoginPost and testLoginPostWithRelayState authnRequestsSigned = True was assumed. Adapt accordingly.

:white_check_mark: Freeze time in time-sensitive tests